`MarkdownDocument` renamed to `Document` and moved into own namespace.
